+++ src/components/Input/style/index.js
@@ -13,7 +13,6 @@ const prefixCls = _prefixCls + '-input';
 export const FixWrap = styled.span`
     vertical-align: middle;
     display: table-cell;
-    color: #636e83;
     /* auto fix fixwrap to min-width */
     width: 1px;
 `;
@@ -31,62 +30,65 @@ export const SearchIcon = styled(Icon)`
 `;
 
 const themeMixin = ({
-    theme: { colorMap, colorList, Height, HeightNumber, materialVars, Input: inputTheme = {} },
+    theme: { designTokens: DT, Height, HeightNumber, materialVars },
     disabled,
     size,
     focused,
     status
 }) => css`
-    color: ${colorList.black};
-    border: 1px solid #dfe0f1;
-    background: #fafafc;
+    border-radius: ${DT.T_CORNER_SM};
     height: ${Height[size]};
-    box-shadow: ${materialVars.innerShadow};
+    color: ${DT.T_COLOR_TEXT_DEFAULT_LIGHT};
+    border: ${DT.T_LINE_WIDTH_BASE} solid ${DT.T_COLOR_LINE_DEFAULT_LIGHT};
+    box-shadow: ${DT.T_SHADOW_INSET_1};
+    background: ${DT.T_INPUT_COLOR_BG_DEFAULT};
     transition: ${materialVars.transitionDown};
     :hover {
-        border-color: #c3cad9;
-        background-color: #f6f6fb;
+        color: ${DT.T_COLOR_TEXT_DEFAULT_DARK};
+        border-color: ${DT.T_COLOR_LINE_DEFAULT_DARK};
+        background: ${DT.T_INPUT_COLOR_BG_DEFAULT};
     }
 
     input {
         line-height: ${HeightNumber[size] - 2}px;
         height: ${HeightNumber[size] - 2}px;
         &::placeholder {
-            color: ${colorList.placeholder};
+            opacity: 1;
+            color: ${DT.T_COLOR_TEXT_REMARK_LIGHT};
         }
     }
     ${focused &&
         !disabled &&
         css`
             && {
-                border-color: ${colorMap.active.border};
-                background-color: #f6f6fb;
+                color: ${DT.T_COLOR_TEXT_DEFAULT_DARK};
+                border-color: ${DT.T_COLOR_LINE_PRIMARY_DEFAULT};
+                background: ${DT.T_INPUT_COLOR_BG_ACTIVE};
             }
         `};
 
     ${status === 'error' &&
         css`
             &&& {
-                border-color: #f44336;
+                border-color: ${DT.T_COLOR_LINE_ERROR_DARK};
+                background: ${DT.T_COLOR_BG_ERROR_LIGHT};
             }
         `};
 
     ${disabled &&
         css`
-            color: ${colorMap.disabled.text};
-            -webkit-text-fill-color: currentcolor;
             box-shadow: none;
             &,
             &:hover {
-                background: ${colorMap.disabled.background};
-                border-color: ${colorMap.disabled.border};
+                color: ${DT.T_COLOR_TEXT_DISABLED};
+                -webkit-text-fill-color: currentcolor;
+                border-color: ${DT.T_COLOR_LINE_DISABLED_DARK};
+                background: ${DT.T_COLOR_BG_DISABLED_LIGHT};
             }
             ${SuffixWrap}, ${PrefixWrap} {
                 color: inherit;
             }
         `};
-
-    ${inputTheme['&']};
 `;
 
 export const TableWrap = styled.span`
@@ -100,7 +102,6 @@ export const InputWrap = styled.span.attrs({
     position: relative;
     ${inlineBlockWithVerticalMixin};
     box-sizing: border-box;
-    border-radius: 2px;
 
     ${({ disabled }) =>
         disabled &&